Move Kebab menu items
